R3 AR-1: move infer_cardinality from parse.rs into graph/
Cardinality inference is semantic-graph logic, not parsing, yet it lived in
parse.rs and was reached back into from the ddl layer
(`crate::parse::infer_cardinality` at ddl/define.rs) — the layering inversion
the review flags under AR-1. Relocate it to `crate::graph::cardinality` and
route all callers (the two CREATE rewrite paths in parse.rs, the enrichment
step in ddl/define.rs, and the phase33 unit tests) through
`crate::graph::infer_cardinality`, so parse and ddl now depend inward on the
graph rather than the reverse. Pure move — no behavior change; the existing
cardinality unit tests exercise it unchanged.

Removes the now-unused `HashSet` and `model::{Cardinality, Join, TableRef}`
imports from parse.rs.
