+++ src/graph/mod.rs
@@ -1,5 +1,6 @@
 //! Relationship graph validation and topological sort for semantic view definitions.
 
+mod cardinality;
 mod derived_metrics;
 mod facts;
 mod names;
@@ -11,6 +12,7 @@ mod using;
 mod test_helpers;
 
 // Public API (matches prior graph.rs surface exactly)
+pub(crate) use cardinality::infer_cardinality;
 pub use derived_metrics::{contains_aggregate_function, validate_derived_metrics};
 pub use facts::{find_fact_references, validate_facts};
 pub use names::validate_name_uniqueness;
